feat: add gui

+++ src/main/java/charginggadgets/utils/IIntArray.java
@@ -1,7 +1,7 @@
 package charginggadgets.utils;
 
 public interface IIntArray {
-    int get(int index);
+    double get(int index);
 
     void set(int index, int value);
